Remove useless spec
This spec simply tests Ruby's inherent exception behaviour. It's
breaking in CI because we're currently running under root user during
builds.

There's an argument to be made that we shouldn't be using root during
test suite runs, and that's being reviewed. Irrespective of this
decision, the spec is still low-value and should be deleted.